pyproject.toml
[project]
name = "comp3030j"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Meow King", email = "mr.meowking@anche.no" }, # ziqi yang 21207440
    { name = "AuroraZoer", email = "yiran.zhao3@ucdconnect.ie" }, # yiran zhao 21207295
    # add yours below
]
dependencies = [
    "flask>=3.0.0",
    "Flask-SQLAlchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "Flask-Babel>=4.0.0",
    "openai==1.14.2",
    "requests>=2.26.0",
    "numpy>=1.21.2",
    "tqdm>=4.62.3",
    "flask_bcrypt>=1.0.1",
    "flask_login>=0.6.3",
    "email_validator>=2.1.1",
    "sqlalchemy_serializer>=1.4.12",
    "SQLAlchemy-Utc>=0.14.0",
    "Flask-Caching>=2.3.0",
    "redis>=5.0.4"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/comp3030j"]
